Moved selection logic out of main
